fix(web, options): wrong argument
