+++ core/repository/src/commonMain/kotlin/com/neoutils/neoregex/core/repository/testcase/TestCasesRepositoryImpl.kt
@@ -68,7 +68,6 @@ internal class TestCasesRepositoryImpl : TestCasesRepository {
     }
 
     override fun clear() {
-
         testCases.update {
             clear()
             put(TestCase())
